requirements.txt
PyPDF2==3.0.1
python-docx==0.8.11
markdown==3.4.3
tiktoken==0.4.0
chardet==5.1.0
requests>=2.26.0
